add validation in ukom-register

+++ src/sijupri-admin/formasi/formasi-document-list/formasi-document-list.component.ts
@@ -1,25 +1,51 @@
+import { ConfirmationService } from './../../../modules/base/services/confirmation.service'
 import { CommonModule } from '@angular/common'
 import { Component } from '@angular/core'
 import { FormsModule } from '@angular/forms'
 import { DokumenPersyaratan } from '../../../modules/maintenance/models/dokumen-persyaratan.model'
 import { ApiService } from '../../../modules/base/services/api.service'
 import { AlertService } from '../../../modules/base/services/alert.service'
-
+import { Pagable } from '../../../modules/base/commons/pagable/pagable'
+import {
+  ActionColumnBuilder,
+  PagableBuilder,
+  PageFilterBuilder,
+  PrimaryColumnBuilder
+} from '../../../modules/base/commons/pagable/pagable-builder'
+import { PagableComponent } from '../../../modules/base/components/pagable/pagable.component'
+import { HandlerService } from '../../../modules/base/services/handler.service'
 @Component({
   selector: 'app-formasi-document-list',
   standalone: true,
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, PagableComponent],
   templateUrl: './formasi-document-list.component.html',
   styleUrl: './formasi-document-list.component.scss'
 })
 export class FormasiDocumentListComponent {
   dokumenPersyaratan: DokumenPersyaratan = new DokumenPersyaratan()
   dokumenPersyaratanList: DokumenPersyaratan[] = []
+  pagable: Pagable
 
   constructor (
     private apiService: ApiService,
-    private alertService: AlertService
-  ) {}
+    private alertService: AlertService,
+    private confirmationService: ConfirmationService,
+    private handlerService: HandlerService
+  ) {
+    this.pagable = new PagableBuilder(
+      '/api/v1/doc_persyaratan/association/for_formasi'
+    )
+      .addPrimaryColumn(new PrimaryColumnBuilder('Nama', 'name').build())
+      .addActionColumn(
+        new ActionColumnBuilder()
+          .setAction((item: any) => {
+            this.delete(item.id)
+          }, 'danger')
+          .withIcon('danger')
+          .build()
+      )
+      .build()
+  }
 
   ngOnInit () {
     this.getDokumentPersyaratanList()
@@ -40,26 +66,45 @@ export class FormasiDocumentListComponent {
   }
 
   submit () {
-    this.dokumenPersyaratan.association = 'for_formasi'
-    this.apiService
-      .postData(`/api/v1/doc_persyaratan`, this.dokumenPersyaratan)
-      .subscribe({
-        next: () => this.getDokumentPersyaratanList(),
-        error: error => {
-          console.error('Error fetching data', error)
-          this.alertService.showToast('Error', error.message)
-          throw error
-        }
-      })
+    this.confirmationService.open(false).subscribe({
+      next: result => {
+        if (!result.confirmed) return
+
+        this.dokumenPersyaratan.association = 'for_formasi'
+
+        this.apiService
+          .postData(`/api/v1/doc_persyaratan`, this.dokumenPersyaratan)
+          .subscribe({
+            next: () => {
+              //   this.getDokumentPersyaratanList()
+              window.location.reload()
+            },
+            error: error => {
+              console.error('Error fetching data', error)
+              this.alertService.showToast('Error', error.message)
+              throw error
+            }
+          })
+      }
+    })
   }
 
   delete (id: string) {
-    this.apiService.deleteData(`/api/v1/doc_persyaratan/${id}`).subscribe({
-      next: () => this.getDokumentPersyaratanList(),
-      error: error => {
-        console.error('Error fetching data', error)
-        this.alertService.showToast('Error', error.message)
-        throw error
+    this.confirmationService.open(false).subscribe({
+      next: result => {
+        if (!result.confirmed) return
+
+        this.apiService.deleteData(`/api/v1/doc_persyaratan/${id}`).subscribe({
+          next: () => {
+            window.location.reload()
+            // this.getDokumentPersyaratanList()
+          },
+          error: error => {
+            console.error('Error fetching data', error)
+            this.alertService.showToast('Error', error.message)
+            throw error
+          }
+        })
       }
     })
   }
